Refactoring modules that will be common to multiple mturk data collections.
