+++ mturk_scripts/singleturn/singleturn_games_storage.py
@@ -1,15 +1,21 @@
+import os
 import random
+import sys
 
 from azure.core.exceptions import ResourceExistsError
-from azure.data.tables import TableServiceClient, TableClient, UpdateMode
-from azure.storage.blob import ContainerClient
 from typing import Any, Dict, List, Optional, Tuple
 
+# Project root
+sys.path.append(os.path.join(os.path.dirname(__file__), '../'))
+
 import logger
+from game_storage import AzureGameStorage
+from turn import Turn
+
 _LOGGER = logger.get_logger(__name__)
 
 
-class Turn:
+class SingleTurnDatasetTurn(Turn):
     """Class that acts as buffer between the database tables and the mturk scripts.
 
     Contains all the data necessary to keep track of a turn, and maps its attributes
@@ -36,26 +42,14 @@ class Turn:
                 world, i.e., previous turn.
         """
         # In singleturn, the game_id also identifies the turn.
-        self.game_id = game_id
-        self.turn_type = turn_type
+        super().__init__(game_id, turn_type, initialized_structure_id)
 
-        self.initialized_structure_id = initialized_structure_id
         self.starting_world_blob_path = starting_world_blob_path
         self.starting_world_blob_name = starting_world_blob_name
         self.starting_step = starting_step
 
         # Path to a sample screenshot of the starting world
         self.screenshot_step_view = starting_step + "_north"
-
-        # Values set after hit has been completed
-        self.hit_id = None
-        self.worker_id = None
-        self.is_qualified = None
-        self.result_blob_path = None
-        self.input_instructions = None
-
-    def set_hit_id(self, hit_id):
-        self.hit_id = hit_id
 
     @staticmethod
     def _add_legacy_keys(row: Dict[str, Any]):
@@ -144,49 +138,13 @@ class Turn:
         return new_entry
 
 
-class IgluSingleTurnGameStorage:
+class IgluSingleTurnGameStorage(AzureGameStorage):
     """Abstraction of storage data structures for single turn games.
 
     This class is a context manager, use inside a with statement.
     >>> with IgluSingleTurnGameStorage("hitTableName", "connectionStr") as game_storage:
     ...     create_new_games(self, starting_structure_ids)
     """
-
-    def __init__(self, hits_table_name: str, azure_connection_str: str,
-                 starting_structures_container_name: str,
-                 starting_structures_blob_prefix: str,
-                 **kwargs) -> None:
-        super().__init__()
-
-        self.azure_connection_str = azure_connection_str
-
-        self.hits_table_name = hits_table_name
-        self.table_client = None
-
-        self.starting_structures_container_name = starting_structures_container_name
-        self.container_client = None
-
-        self.starting_structures_blob_prefix = starting_structures_blob_prefix
-        self.blob_service_client = None
-
-    def __enter__(self):
-        with TableServiceClient.from_connection_string(
-                self.azure_connection_str) as table_service_client:
-            _ = table_service_client.create_table_if_not_exists(
-                table_name=self.hits_table_name)
-
-        self.table_client = TableClient.from_connection_string(
-            conn_str=self.azure_connection_str, table_name=self.hits_table_name)
-        self.table_client.__enter__()
-        _LOGGER.debug(f"Entering {self.__class__.__name__} context anc closing TableClient.")
-        return self
-
-    def __exit__(self, exception_type, exception_value, traceback):
-        self.table_client.__exit__(
-            exception_type, exception_value, traceback)
-        _LOGGER.debug(f"Leaving {self.__class__.__name__} context and closing TableClient.")
-        self.table_client = None
-        return None
 
     def get_last_game_index(self, turn_type: str = 'builder-normal') -> int:
         """Returns the maximum game index, stored in column PartitionKey of the hits table.
@@ -254,9 +212,7 @@ class IgluSingleTurnGameStorage:
         if game_count <= 0:
             raise ValueError("Not creating any Hits as games count is less than 1")
 
-        container_client = ContainerClient.from_connection_string(
-            self.azure_connection_str,
-            self.starting_structures_container_name)
+        container_client = self.create_container_client()
         blob_names = container_client.list_blobs(
             name_starts_with=self.starting_structures_blob_prefix)
         blob_list = []
@@ -287,7 +243,7 @@ class IgluSingleTurnGameStorage:
             initialized_structure_id = starting_world_blob_name.split('-')[1]
         else:
             initialized_structure_id = starting_world_blob_name
-        new_turn = Turn(
+        new_turn = SingleTurnDatasetTurn(
             game_id, turn_type,
             initialized_structure_id=initialized_structure_id,
             starting_world_blob_path=starting_world_blob_path,
@@ -336,24 +292,3 @@ class IgluSingleTurnGameStorage:
 
         except (ResourceExistsError, StopIteration):
             _LOGGER.warning(f'No turn with {column_name} = {key} found on table')
-
-    def save_new_turn(self, turn: Turn):
-        if turn.hit_id is None:
-            _LOGGER.warning(f"Attempting to save turn without created hit for game {turn.game_id}")
-            return
-
-        entity = turn.to_database_entry(self.starting_structures_container_name)
-
-        try:
-            self.table_client.create_entity(entity)
-            _LOGGER.debug(f'Successfully inserted new turn {turn.hit_id} for game {turn.game_id}.')
-        except ResourceExistsError:
-            _LOGGER.error(f"Turn entry {turn.hit_id} for game {turn.game_id} already exists")
-
-    def upsert_turn(self, turn: Turn):
-        if turn.hit_id is None:
-            _LOGGER.warning(
-                f"Attempting to upsert turn without created hit for game {turn.game_id}")
-            return
-        entity = turn.to_database_entry(self.starting_structures_container_name)
-        self.table_client.upsert_entity(mode=UpdateMode.MERGE, entity=entity)
