+++ mturk_scripts/singleturn/tests/ping_azure_tables.py
@@ -9,10 +9,12 @@ import dotenv
 import os
 import sys
 
-sys.path.append(os.path.join(os.path.dirname(__file__), '../'))
+# Project root
+sys.path.append(os.path.join(os.path.dirname(__file__), '../../'))
 
 import logger  # noqa: E402
-from singleturn_games_storage import IgluSingleTurnGameStorage, Turn  # noqa: E402
+from singleturn.singleturn_games_storage import (
+    IgluSingleTurnGameStorage, SingleTurnDatasetTurn)  # noqa: E402
 from utils import read_config  # noqa: E402
 
 # Load dotenv before project imports
@@ -36,7 +38,7 @@ def read_args():
 def main():
     args = read_args()
     connection_str = os.getenv('AZURE_STORAGE_CONNECTION_STRING')
-    config = read_config(args.config, '../env_configs.json')
+    config = read_config(args.config, config_filepath='../env_configs.json')
 
     turn_type = 'builder-normal'
     with IgluSingleTurnGameStorage(
@@ -48,7 +50,7 @@ def main():
 
         _LOGGER.info("Random turn row successfully retrieved")
 
-        new_turn = Turn.from_database_entry(random_row)
+        new_turn = SingleTurnDatasetTurn.from_database_entry(random_row)
         generated_row = new_turn.to_database_entry(game_storage.starting_structures_container_name)
 
         _LOGGER.info("Turn created from database row")
@@ -75,7 +77,7 @@ def main():
             retrieved_row = game_storage.retrieve_turn_entity(
                 key=new_turn.hit_id, column_name='RowKey', qualified_value=True)
 
-            recreated_turn = Turn.from_database_entry(retrieved_row)
+            recreated_turn = SingleTurnDatasetTurn.from_database_entry(retrieved_row)
             if recreated_turn.input_instructions != new_turn.input_instructions:
                 _LOGGER.error(
                     f"Upserting entity did not replace field. Expected: "
