+++ mturk_scripts/tests/test_data/simple_template.xml
