+++ mturk_scripts/tests/test_hit_manager.py
@@ -11,7 +11,7 @@ from unittest import mock
 
 sys.path.append(os.path.join(os.path.dirname(__file__), '../'))
 
-from hit_manager import HITManager, BuilderTemplateRenderer  # noqa: E402
+from hit_manager import HITManager  # noqa: E402
 
 
 class MturkClientFake(mock.MagicMock):
@@ -138,36 +138,6 @@ class HitManagerTest(unittest.TestCase):
 
         hit_ids = hit_manager.get_open_hit_ids(self.HIT_TYPE)
         self.assertEqual(set(hit_ids), set(expected_hit_ids))
-
-
-class BuilderTemplateRendererTests(unittest.TestCase):
-
-    def test_render_builder_template(self):
-        """Builder normal template is correctly rendered.
-
-        If this tests fails, it is likely there is a discrepancy between function
-        builder_normal_template_kwargs and the template itself."""
-
-        renderer = BuilderTemplateRenderer(
-            template_filepath="test_data/no_write_builder_normal.xml",
-        )
-
-        template_kwargs = {
-            'game_id': 'game_id',
-            'azure_sas': 'azure_sas',
-            'starting_world_blob_path': 'test-builder-data',
-            'starting_world_blob_name': '37-c161',
-            'starting_step': 'step-2',
-            'screenshot_step_view': 'step-2_north',
-        }
-
-        rendered_template = renderer.render_template(**template_kwargs)
-
-        # All variables have been replaced
-        self.assertNotIn('$', rendered_template)
-
-        for parameter_value in template_kwargs.values():
-            self.assertIn(parameter_value, rendered_template)
 
 
 if __name__ == '__main__':
